requirements.txt
pyyaml
twine
wheel
setuptools